- primary key ddl was broken; fixed the if not exists queries - added a couple if not exists to the index ddl - for testing no rxnorm extension from server, added   delete_rxnorm_extension_records function (in initialize)   so if you call initialize with optimization_experiment='n3c_no_rxnorm'   it will run the delete; make sure to set the schema to something where   you actually want to delete those records. we probably won't have   to use it again as it is, but may want to try other experiments using   similar functions, so leaving in place. - while working on it had two significant problems:   - load_csv kept crashing in the middle of uploading     concept_relationship   - and then I accidentally dropped concept_relationship in the n3c     schema and had the same issue trying to restore that   - to fix it I added 'finish aborted upload' as a replace_rule option     to load_csv, and added replace_rule to seed and initialize so you     can send that instead of just clobber true/false   - I also added chunksize=1000 to the df.to_sql call   - I didn't have (or need) dowloaded data for omopconceptset and     omopconceptsetcontainer, so load_csv crashed when expecting those     files and not finding them. fixed that.

+++ backend/db/utils.py
@@ -513,7 +513,10 @@ def load_csv(
     schema: str = SCHEMA, is_test_table=False, local=False, optional_suffix=''
 ):
     """Load CSV into table
-    :param replace_rule: 'replace if diff row count' or 'do not replace'
+    :param replace_rule:
+        - 'replace if diff row count'
+        - 'do not replace'
+        - 'finish aborted upload'
       First, will replace table (that is, truncate and load records; will fail if table cols have changed, i think
      'do not replace'  will create new table or load table if table exists but is empty
     :param optional_suffix: Useful for when remaking tables when database is live. For example, you can upload a new
@@ -543,6 +546,10 @@ def load_csv(
     # Load table
     path = os.path.join(DATASETS_PATH, f'{table_name_no_suffix}.csv') if table_type == 'dataset' \
         else os.path.join(OBJECTS_PATH, table, 'latest.csv')
+    if not os.path.isfile(path):
+        print(f'INFO: {path} does not exist; skipping')
+        return
+
     df = pd.read_csv(path)
 
     if is_test_table:
@@ -552,18 +559,22 @@ def load_csv(
         print(f'INFO: {schema}.{table} exists with same number of rows {existing_rows}; leaving it')
         return
 
-    print(f'INFO: \nloading {schema}.{table} into {CONFIG["server"]}:{DB}')
-    # Clear data if exists
-    try:
-        con.execute(text(f'DROP TABLE {schema}.{table} CASCADE'))
-    except ProgrammingError:
-        pass
+    if replace_rule == 'finish aborted upload' and existing_rows < len(df):
+        print(f'INFO: {schema}.{table} exists with {commify(existing_rows)} rows; uploading remaining {commify(len(df)-existing_rows)} rows');
+        df = df.iloc[existing_rows: len(df)]
+    else:
+        print(f'INFO: \nloading {schema}.{table} into {CONFIG["server"]}:{DB}')
+        # Clear data if exists
+        try:
+            con.execute(text(f'DROP TABLE {schema}.{table} CASCADE'))
+        except ProgrammingError:
+            pass
 
     # Load
     # `schema='termhub_n3c'`: Passed so Joe doesn't get OperationalError('(pymysql.err.OperationalError) (1050,
     #  "Table \'code_sets\' already exists")')
     #  https://stackoverflow.com/questions/69906698/pandas-to-sql-gives-table-already-exists-error-with-if-exists-append
-    kwargs = {'if_exists': 'append', 'index': False, 'schema': schema}
+    kwargs = {'if_exists': 'append', 'index': False, 'schema': schema, 'chunksize': 1000}
     # TODO: add suffix
     df.to_sql(table, con, **kwargs)
